requirements.txt
aiogram<3.0
aiohttp==3.6.2
aioredis<2.0
python-dotenv~=0.19.1
Django~=3.2.8
psycopg2-binary
gunicorn